# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Science AI Writer IDE — API locale
  version: 0.3.1
  description: |
    Contrat de l'API v1 du Science AI Writer IDE.

    Ce fichier est normatif : il est subordonné aux ADR et aux spécifications
    techniques V0.3, et prioritaire sur toute implémentation existante.
    Une divergence entre le code et ce contrat est un défaut du code.

    Principes :
    - API strictement locale, liée à 127.0.0.1. Aucune exposition réseau.
    - Les opérations longues (ingestion, génération, export) répondent 202 et
      retournent une tâche ; la progression est suivie par SSE.
    - Les portes de validation humaine sont des endpoints explicites. Aucune
      transition d'état ne les franchit automatiquement.
    - Toute sortie réseau exige un consentement de périmètre préalable.
  license:
    name: Propriétaire
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Instance locale
...
paths:
...
  /projects/{projectId}/code/propose:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [code]
      summary: Faire produire une analyse par l'agent
      description: |
        L'agent code écrit le script, le serveur l'exécute au niveau Wasm
        (ADR-005), puis rapproche les artefacts attendus des artefacts
        réellement produits. Un attendu absent relance l'agent avec un message
        précis ; un produit non déclaré est conservé et signalé, jamais
        rattaché seul. La boucle de correction est plafonnée (US-202), et une
        indisponibilité de paquet ne relance pas l'agent : elle remonte avec la
        proposition du niveau natif, qui est une décision humaine.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intent]
              properties:
                intent: { type: string }
                datasets:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Exécution retenue et ses artefacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CodeRun' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /projects/{projectId}/code/execute:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [code]
      summary: Exécuter un script Python en environnement isolé
      description: |
        Le niveau d'isolation est décidé par le serveur (ADR-005) :
        toute exécution d'origine agent est ramenée au niveau Wasm, quel que
        soit le mode demandé. Le niveau natif exige le consentement
        native_execution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, origin]
              properties:
                code: { type: string }
                origin: { type: string, enum: [agent, user] }
                mode: { type: string, enum: [wasm, native], default: wasm }
                limits: { $ref: '#/components/schemas/ResourceLimits' }
                datasets:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Résultat d'exécution
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionResult' }
        '403': { $ref: '#/components/responses/ConsentRequired' }
  /projects/{projectId}/code/executions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [code]
      summary: Historique des exécutions
      responses:
        '200':
          description: Historique
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExecutionResult' }
  /projects/{projectId}/sections/{sectionId}/artifacts/{artifactId}/attach:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/SectionId'
      - $ref: '#/components/parameters/ArtifactId'
    post:
      tags: [code]
      summary: Rattacher un artefact à une section
      description: |
        Insère le renvoi Quarto dans content_qmd s'il n'y figure pas déjà —
        ![légende](chemin){#label} pour une figure (ADR-006). L'opération est
        IDEMPOTENTE : un artefact déjà référencé n'ajoute pas un second renvoi.
        Seuls une figure et un tableau se rattachent ; un artefact de données
        n'a pas de renvoi dans le texte.
      responses:
        '200':
          description: Section mise à jour, renvoi inséré
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftSection' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /projects/{projectId}/artifacts/{artifactId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/ArtifactId'
    delete:
      tags: [code]
      summary: Supprimer un artefact
      description: |
        Refusée avec 409 si l'artefact est référencé dans une section : le
        supprimer laisserait un renvoi sans cible et ferait échouer la
        compilation des semaines plus tard (US-502).
      responses:
        '204': { description: Artefact supprimé }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Artefact référencé dans une section
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: integer }
...
    SectionId:
      name: sectionId
      in: path
      required: true
      schema: { type: integer }
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: integer }
  responses:
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Corps de requête invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ConsentRequired:
      description: Consentement de périmètre absent
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConsentRequiredDetail' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    ConsentRequiredDetail:
      type: object
      required: [code, scope, message]
      properties:
        code: { type: string, const: CONSENT_REQUIRED }
        scope: { $ref: '#/components/schemas/ConsentScope' }
        message: { type: string }
        will_transmit:
          type: string
          description: Description de ce qui sortirait si le consentement était accordé.
...
    DraftSection:
      type: object
      required: [id, plan_node_id, content_qmd, status, version]
      properties:
        id: { type: integer }
        plan_node_id: { type: integer }
        content_qmd:
          type: string
          description: Quarto Markdown, format canonique unique (ADR-006).
        status:
          type: string
          enum: [DRAFTING, GUARDRAIL, REVIEWING, CORRECTING, VALIDATED, ORPHANED]
        quality_score: { type: number, nullable: true, minimum: 0, maximum: 100 }
        version: { type: integer }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        generated_at: { type: string, format: date-time, nullable: true }
        validated_at: { type: string, format: date-time, nullable: true }
    Citation:
      type: object
      required: [id, source_id, bibtex_key, verified]
      properties:
        id: { type: integer }
        source_id: { type: integer }
        chunk_id: { type: integer, nullable: true }
        bibtex_key: { type: string }
        locator: { type: string, nullable: true }
        verified:
          type: boolean
          description: Une citation non vérifiée bloque l'export (ADR-007).
    ResourceLimits:
      type: object
      properties:
        memory_mb: { type: integer, default: 2048 }
        cpu_seconds: { type: integer, default: 60 }
        wall_seconds: { type: integer, default: 120 }
        max_file_mb: { type: integer, default: 512 }
        max_processes: { type: integer, default: 8 }
    ExecutionResult:
      type: object
      required: [exit_code, stdout, stderr, duration_ms, level, network_isolation_guaranteed]
      properties:
        id: { type: integer }
        exit_code: { type: integer }
        stdout: { type: string }
        stderr: { type: string }
        duration_ms: { type: integer }
        level: { type: integer, enum: [1, 2] }
        timed_out: { type: boolean }
        limit_exceeded:
          type: string
          nullable: true
          enum: [memory, cpu, wall, null]
        network_isolation_guaranteed:
          type: boolean
          description: |
            Faux au niveau 2 sous Windows. Ce champ est honnête par
            construction et affiché à l'utilisateur (ADR-005).
        downgraded_from_requested_mode: { type: boolean }
        artifacts:
          type: array
          items: { type: string }
    Artifact:
      type: object
      description: |
        Fichier produit par une exécution, porteur de quoi le REPRODUIRE :
        code exact, graine, versions des bibliothèques relevées dans le bac à
        sable, SHA-256 des jeux de données (US-401). Sans le SHA-256, la
        question « comment cette figure a-t-elle été obtenue » reste sans
        réponse complète : un fichier modifié depuis en produirait une autre.
      required: [id, code_execution_id, kind, filename, rel_path, declared]
      properties:
        id: { type: integer }
        code_execution_id: { type: integer }
        draft_section_id: { type: integer, nullable: true }
        kind: { type: string, enum: [figure, table, data] }
        filename: { type: string }
        label:
          type: string
          nullable: true
          description: Renvoi Quarto — fig-<slug> ou tbl-<slug>. Jamais du MyST.
        caption: { type: string, nullable: true }
        rel_path:
          type: string
          description: Chemin relatif au répertoire de données, jamais absolu.
        declared:
          type: boolean
          description: |
            Faux pour un artefact produit mais NON déclaré : conservé et
            signalé, jamais rattaché automatiquement à une section.
        random_seed: { type: integer, nullable: true }
        library_versions:
          type: object
          additionalProperties: { type: string }
        dataset_sha256:
          type: object
          additionalProperties: { type: string }
        duration_ms: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    CodeRun:
      type: object
      description: |
        Exécution retenue d'une proposition d'agent, et ses artefacts (US-401).
        Distincte d'`ExecutionResult`, qui décrit une exécution directe sans
        rapprochement d'artefacts.
      required: [id, origin, sandbox_level, exit_code, duration_ms, started_at]
      properties:
        id: { type: integer }
        origin: { type: string, enum: [agent, user] }
        sandbox_level: { type: string, enum: [wasm, native] }
        exit_code: { type: integer }
        duration_ms: { type: integer }
        started_at: { type: string, format: date-time }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/Artifact' }
...
    ConsentScope:
      type: string
      enum:
        - biblio_search
        - model_download
        - plagiarism_check
        - remote_llm
        - native_execution
...